Move property query fragments into own folder
