+++ src/core/database/query/index.ts
@@ -1,17 +1,17 @@
 export * from './create-node';
-export * from './create-property';
+export * from './properties/create-property';
+export * from './properties/deactivate-property';
+export * from './properties/update-property';
+export * from './properties/update-properties';
 export * from './create-relationships';
 export * from './cypher-expression';
 export * from './cypher-functions';
 export * from './full-text';
-export * from './deactivate-property';
 export * from './lists';
 export * from './matching';
 export * from './deletes';
 export * from './match-project-based-props';
 export * from './match-changeset-and-changed-props';
-export * from './update-property';
-export * from './update-properties';
 export { QueryFragment } from '../query-augmentation/apply';
 export * from '../query-augmentation/condition-variables';
 export * from './where-path';
